Meta: Added `TTTReborn.Globals`, including `Utils` and `RPCs` (#98)
