+++ code/player/TTTPlayer.Commands.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using Sandbox;
 
+using TTTReborn.Globals;
 using TTTReborn.Items;
 using TTTReborn.Roles;
 
@@ -131,7 +132,7 @@ namespace TTTReborn.Player
             }
 
             player.SetRole(role);
-            ClientSetRole(To.Single(player), player, role.Name);
+            RPCs.ClientSetRole(To.Single(player), player, role.Name);
         }
 
         [ClientCmd(Name = "playerids", Help = "Returns a list of all players (clients) and their associated IDs")]
