+++ code/player/TTTPlayer.Roles.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 
 using Sandbox;
 
+using TTTReborn.Globals;
 using TTTReborn.Roles;
 using TTTReborn.Teams;
 
@@ -84,11 +85,11 @@ namespace TTTReborn.Player
                     }
                 }
 
-                ClientAddMissingInAction(To.Multiple(traitors), this);
+                RPCs.ClientAddMissingInAction(To.Multiple(traitors), this);
             }
             else
             {
-                ClientAddMissingInAction(To.Single(player), this);
+                RPCs.ClientAddMissingInAction(To.Single(player), this);
             }
         }
     }
