+++ code/ui/ChatBox.cs
@@ -4,6 +4,7 @@ using Sandbox;
 using Sandbox.UI;
 using Sandbox.UI.Construct;
 
+using TTTReborn.Globals;
 using TTTReborn.Player;
 
 // Chat needs to hook into s&box in order to function.
