+++ code/ui/Scoreboard.cs
@@ -4,6 +4,7 @@ using Sandbox;
 using Sandbox.UI;
 using Sandbox.UI.Construct;
 
+using TTTReborn.Globals;
 using TTTReborn.Player;
 
 namespace TTTReborn.UI
